New Features + Refactoring
Project structure changed to the following:
Director: Contains files meant to directly change/interact with data library
API: Empty for now, meant for scripts that extract data from library and do some processing on it
InputFiles: If new information is to be added to the library these files go in here, they can be promptly removed once parsing is finished

New Feature:
It is now possible to edit the asdf files using the director.
Adding and removing datasets directly from already created files can be done.
Future: Edit bib information (currently bugged) and adding notes to the file

This is the last entry for the repository,  Follow
https://github.com/openmsr/msdl
for any future updates